+++ lib/src/extensions/date_time_extension/_time_ago_bucket.dart
@@ -1,0 +1,12 @@
+part of 'date_time_extension.dart';
+
+enum _TimeAgoBucket {
+  years,
+  months,
+  days,
+  hours,
+  minutes,
+  seconds,
+  milliseconds,
+  now,
+}